Use a better approach to capturing the main functions exported from the translator
The translators create several top-level functions which
need to be invoked externally.

Previously this was done by rewriting 'function <name>' to 'export <name> = function',
however this broke calls to the function elsewhere in the translator.

This commit uses a simpler approach of setting exports.<name> at the bottom
of the file.