Work towards supporting the Google Books translator
 * Implement Zotero.doGet() API
 * Implement DOMParser() API

The Google Books translator is currently not working
because the translator assumes that all items have been created via 'new
Zotero.Item()' by the time that doWeb() returns.

This is not the case however for the Google Books importer as it
initiats async fetches via Zotero.doGet() and only creates items
once it has parsed the results.