Implement support for Google Books importer
Refactor translator API to return an RX.js observable
instead of a promise. This is needed because the translator
may produce items asychronously, as is the case with Google Books
where the page is parsed and then additional network requests
are issued to fetch metadata for books in XML format.

Translator.processPage() now returns an observable which is closed
once the translator has run and the count of active network
requests triggered by the translator reaches zero.